requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.1
python-multipart==0.0.9
httpx==0.27.0

# Streamlit Frontend
streamlit==1.35.0
requests==2.32.3

# Supabase
supabase==2.5.0

# LangChain + RAG
langchain==0.2.3
langchain-community==0.2.3
langchain-huggingface==0.0.3
faiss-cpu==1.8.0

# Hugging Face
huggingface-hub==0.23.2
sentence-transformers==3.0.0
datasets==2.20.0
transformers==4.41.2
torch==2.3.0

# File Processing
PyMuPDF==1.24.5
pandas==2.2.2
openpyxl==3.1.4
python-docx==1.1.2

# Utilities
tiktoken==0.7.0
numpy==1.26.4
